+++ src/payloads.rs
@@ -178,3 +178,9 @@ pub use stop_poll::{StopPoll, StopPollSetters};
 pub use unban_chat_member::{UnbanChatMember, UnbanChatMemberSetters};
 pub use unpin_chat_message::{UnpinChatMessage, UnpinChatMessageSetters};
 pub use upload_sticker_file::{UploadStickerFile, UploadStickerFileSetters};
+
+// end of auto generated block
+
+mod get_updates_fault_tolerant;
+
+pub use get_updates_fault_tolerant::GetUpdatesFaultTolerant;